refactor fetching and util function
